+++ libs/common/src/lib/base/index.ts
@@ -1,3 +1,4 @@
 export * from './base.component';
 export * from './base.module';
+export * from './dashboard-layout/dashboard-layout.component';
 export * from './empty.component';